chore(app): change to support only undo
